requirements.txt
aiogram>=3.0.0
fastapi
uvicorn[standard]
asyncpg
python-dotenv
jinja2
python-multipart
redis